2019-08-07提交说明: 1.表增加is_deleted表示 2.优化字典功能 3.修改页面重置按钮优化

+++ aquaman-security-os/src/main/java/com/aquaman/security/admin/entity/domain/UploadFileInfo.java
@@ -61,9 +61,9 @@ public class UploadFileInfo implements Serializable {
     private Long originSystemId;
 
     /**
-     * (0:删除, 1:未删除)
+     * 是否删除(0:否;1:是)
      */
-    private Long isDeleted;
+    private Integer isDeleted;
 
     private Long createUserId;
